debian packaging: swith to libcurl4-openssl
Previously, gnutls variant was used, but this is incompatible with
librdkafka on newer debian. librdkafka uses openssl variant, and both
variants of libcurl cannot coexist in the same system.

As such, moved to openssl in rsyslog curl handling as well. This shall
not affect user interface or functionality in any way.